Move TextLayers to a CompositeLayer
