+++ src/components/heatmap/utils.js
@@ -1,18 +1,11 @@
 import clamp from 'lodash/clamp';
 import range from 'lodash/range';
 
-export const COLOR_BAR_SIZE = 20;
-export const LOADING_TEXT_SIZE = 14;
-export const AXIS_LABEL_TEXT_SIZE = 9;
-export const AXIS_TITLE_TEXT_SIZE = 15;
-export const AXIS_MIN_SIZE = 10;
-export const AXIS_MAX_SIZE = 90;
-export const AXIS_MARGIN = 3;
-export const THEME_TO_TEXT_COLOR = {
-  dark: [224, 224, 224],
-  light: [64, 64, 64],
-};
-export const AXIS_FONT_FAMILY = "-apple-system, 'Helvetica Neue', Arial, sans-serif";
+import {
+  AXIS_LABEL_TEXT_SIZE,
+  AXIS_MIN_SIZE,
+  AXIS_MAX_SIZE,
+} from '../../layers/heatmap-constants';
 
 export function getGeneByCellTile(view, {
   tileSize, tileI, tileJ, numCells, numGenes, cellOrdering, cells,
